Move flatten_dict to new writer utils module
